requirements.txt
gymnasium[mujoco]>=1.1.1
tensorflow==2.17.0
tensorflow-probability==0.22.1
tensorstore==0.1.65
tf-keras==2.17.0
numpy==1.26.4
tensorboard>=2.17.1
tqdm>=4.66.5 
jax==0.4.30
optax>=0.1.9
flax>=0.9.0
ml-dtypes>=0.4.1
ruamel.yaml<0.18.0